bot/requirements.txt
aiogram==3.2.0
aiohttp==3.9.1
httpx==0.25.2
python-dotenv==1.0.0